requirements.txt
Jinja2==3.1.2
bibtexparser==1.4.0
black==22.12.0
colorama==0.4.6
dulwich==0.20.50
ipykernel==6.19.4
ipython==8.7.0
linkify-it-py==2.0.0
markdown-it-py==2.1.0
mdit-py-plugins==0.3.3
nbconvert==7.2.7
pyyaml==6.0.0
